requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2
django-redis==5.4.0
dotenv==0.9.9
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
kavenegar==1.1.2
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0